requirements.txt
certifi==2019.6.16
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
Django==2.2.4
django-cors-headers==3.0.2
django-rest-swagger==2.2.0
django-summernote==0.8.11.4
django-templated-mail==1.1.1
djangorestframework==3.10.2
djangorestframework-jsonapi==2.8.0
djangorestframework-jwt==1.11.0
djoser==2.0.1
idna==2.8
inflection==0.3.1
itypes==1.1.0
Jinja2==2.10.1
MarkupSafe==1.1.1
openapi-codec==1.3.2
pkg-resources==0.0.0
postgres==2.2.2
psycopg2-binary==2.8.3
PyJWT==1.7.1
pytz==2019.2
requests==2.22.0
simplejson==3.16.0
six==1.12.0
sqlparse==0.3.0
uritemplate==3.0.0
urllib3==1.25.3
